remove Spark 2.0 support
