+++ .travis.yml
@@ -12,16 +12,12 @@ env:
 
 matrix:
   include:
-    - jdk: openjdk7 # openJDK crashes sometimes
-      env: PYSPARK_PYTHON=python2 SPARK_VERSION=2.0.2 SPARK_BUILD="spark-${SPARK_VERSION}-bin-hadoop2.7"
     - jdk: oraclejdk8
       env: PYSPARK_PYTHON=python2 SPARK_VERSION=2.1.2 SPARK_BUILD="spark-${SPARK_VERSION}-bin-hadoop2.7"
     - jdk: oraclejdk8
       env: PYSPARK_PYTHON=python2 SPARK_VERSION=2.2.1 SPARK_BUILD="spark-${SPARK_VERSION}-bin-hadoop2.7"
     - jdk: oraclejdk8
       env: PYSPARK_PYTHON=python2 SPARK_VERSION=2.3.0 SPARK_BUILD="spark-${SPARK_VERSION}-bin-hadoop2.7"
-    - jdk: openjdk7
-      env: PYSPARK_PYTHON=python3 SPARK_VERSION=2.0.2 SPARK_BUILD="spark-${SPARK_VERSION}-bin-hadoop2.7"
     - jdk: oraclejdk8
       env: PYSPARK_PYTHON=python3 SPARK_VERSION=2.1.2 SPARK_BUILD="spark-${SPARK_VERSION}-bin-hadoop2.7"
     - jdk: oraclejdk8
