+++ Makefile
@@ -1,9 +1,9 @@
-all: 2.2.0s2.10 2.0.2 2.1.1 2.2.0 2.3.0
+all: 2.2.0s2.10 2.1.1 2.2.0 2.3.0
 
 clean:
 	rm -rf target/graphframes_*.zip
 
-2.0.2 2.1.1 2.2.0 2.3.0:
+2.1.1 2.2.0 2.3.0:
 	build/sbt -Dspark.version=$@ spDist
 
 2.2.0s2.10:
