+++ build.sbt
@@ -4,7 +4,6 @@
 val sparkVer = sys.props.getOrElse("spark.version", "2.3.0")
 val sparkBranch = sparkVer.substring(0, 3)
 val defaultScalaVer = sparkBranch match {
-  case "2.0" => "2.11.8"
   case "2.1" => "2.11.8"
   case "2.2" => "2.11.8"
   case "2.3" => "2.11.8"
@@ -51,14 +50,6 @@ libraryDependencies ++= Seq(
 )
 
 parallelExecution := false
-
-unmanagedSourceDirectories in Compile ++=
-  Seq(baseDirectory.value / "src" / "main" / {
-    sparkBranch match {
-      case ver if ver.startsWith("2.0") => "spark-2.0"
-      case _ => "spark-2.x"
-    }
-  })
 
 scalacOptions in (Compile, doc) ++= Seq(
   "-groups",
